+++ tests/framework/filters/CorsTest.php
@@ -64,4 +64,24 @@ class CorsTest extends TestCase
         $this->assertEquals('*', $cors->response->getHeaders()->get('access-control-allow-origin'));
     }
 
+    public function testAccessControlAllowHeadersPreflight() {
+        $this->mockWebApplication();
+        $controller = new Controller('id', Yii::$app);
+        $action = new Action('test', $controller);
+        $request = new Request();
+        $cors = new Cors([
+            'cors' => [
+                'Origin' => ['*',],
+                'Access-Control-Allow-Headers' => ['authorization','X-Requested-With','content-type', 'custom_header']
+            ],
+        ]);
+        $cors->request = $request;
+
+        $_SERVER['REQUEST_METHOD'] = 'GET';
+        $_SERVER['HTTP_ORIGIN'] = 'http://foo.com';
+        $this->assertTrue($cors->beforeAction($action));
+        $this->assertEquals('authorization, X-Requested-With, content-type, custom_header', $cors->response->getHeaders()->get('Access-Control-Allow-Headers'));
+    }
+
+
 }